requirements.txt
requests
ffmpeg-python
twitchAPI
asyncio
flask
waitress
python-dotenv
PyLivestream
pyjwt